WI #2735 Refactor completion after OF
